Common compilation fixes
xrGame, xrAICore, xrNetServer, xrSound

+++ src/xrAICore/Navigation/graph_abstract.h
@@ -34,7 +34,7 @@ public:
 
 private:
     VERTICES m_vertices;
-    u32 m_edge_count;
+    size_t m_edge_count;
 
 public:
     IC CGraphAbstract();
